% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{GetWeekPredCI}
\alias{GetWeekPredCI}
\title{Calculate median recruitment with 95\% CI for the next 52 weeks}
\usage{
GetWeekPredCI(nSim = 10000L, fillGaps = FALSE, cauchyWt = FALSE, coeff = 1)
}
\arguments{
\item{nSim}{Number of simulations to run (default = 1e4)}

\item{fillGaps}{Whether to fill gaps in the data (default = FALSE)}

\item{cauchyWt}{Whether to use Cauchy weights for sampling (default = FALSE).
If FALSE, binomial weights are used.}

\item{coeff}{A coefficient to apply to the recruitment rate (default = 1)}
}
\value{
An 52x3 matrix with the 2.5\%, 50\% and 97.5\% percentiles for each week
}
\description{
Calculate median recruitment with 95\% CI for the next 52 weeks
}
\examples{
LoadData(gripsYR1, ScreenDt, Enrolled)
(res <- GetWeekPredCI(fillGaps = TRUE, coeff = 1.5))
scenarios <- list(
  sc1 = GetWeekPredCI(),
  sc2 = GetWeekPredCI(cauchyWt = TRUE),
  sc3 = GetWeekPredCI(fillGaps = TRUE),
  sc4 = GetWeekPredCI(fillGaps = TRUE, coeff = 1.2)
)
maxY <- sapply(scenarios, \(x) x$plot_$maxY) |> max()

graphics::par(mfrow = c(2, 2))
for (x in scenarios) x$predPlot(yMax = maxY, Title = x$call.)
}
\seealso{
Other RCTRecruit functions: 
\code{\link{GetDistance}()},
\code{\link{LoadData}()},
\code{\link{Time2Nsubjects}()}
}
\concept{RCTRecruit functions}
% f4
