% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getWeeksPredCI}
\alias{getWeeksPredCI}
\title{Function: Calculate median recruitment with CI for the next 52 weeks}
\usage{
getWeeksPredCI(nSim = 10000L, fillGaps = FALSE, cauchyWt = FALSE, coeff = 1)
}
\arguments{
\item{nSim}{Number of simulations to run}

\item{fillGaps}{Whether to fill gaps in the data}

\item{cauchyWt}{Whether to use Cauchy weights for sampling (default = FALSE).
If FALSE, binomial weights are used.}

\item{coeff}{A coefficient to apply to the recruitment rate (default = 1)}
}
\value{
An 52x3 matrix with the 2.5\%, 50\% and 97.5\% percentiles for each week
}
\description{
Function: Calculate median recruitment with CI for the next 52 weeks
}
\examples{
LoadData(gripsIM, ScreenDt, Enrolled)
getWeeksPredCI()
}
