% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripsIM.R
\docType{data}
\name{gripsIM}
\alias{gripsIM}
\title{Data: Recruitment data for the GRIPS study}
\format{
A data frame with
\strong{\verb{r nrow(dat)}} observations of
\strong{\verb{r length(dat)}} variables:
\verb{r genDataDoc(dat)}
}
\source{
\verb{r addRef("Villasante2024")}
}
\usage{
gripsIM
}
\description{
Modified recruitment data from the GRIPS study for the first year the study
was conducted. The data includes the number of participants recruited each
for each calendar date the recruitment was active.
}
\references{
\enumerate{
\item \verb{r addRef("Villasante2024")}
\item \verb{r addRef("Gajewski2008")}
\item \verb{r addRef("Jiang2015")}
}
}
\keyword{datasets}
