% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gripsYR2Weekly}
\alias{gripsYR2Weekly}
\title{Weekly recruitment data for the 2nd year of the GRIPS study}
\format{
A data frame with \strong{52} observations of \strong{4} variables\tabular{llll}{
    \tab  \tab  \tab  \cr
   [,1] \tab week \tab \code{double} \tab Calendar week \cr
   [,2] \tab year \tab \code{double} \tab Calendar year \cr
   [,3] \tab enrolled \tab \code{integer} \tab Number of people enrolled that week \cr
   [,4] \tab activeDays \tab \code{integer} \tab Number of days in that week when recruitment was active \cr
}
}
\source{
Villasante-Tezanos A, Kuo Y, Kurinec C, Li Y, Yu X (2024). "A
non-parametric approach to predict the recruitment for randomized
clinical trials: an example in elderly inpatient settings." \emph{BMC
medical research methodology}, \emph{24}, 189. ISSN 1471-2288,
\url{https://pubmed.ncbi.nlm.nih.gov/39210285/}.
}
\usage{
gripsYR2Weekly
}
\description{
Modified recruitment data from the second year of the GRIPS study,
aggregated by calendar week.
}
\seealso{
Other Links: 
\code{\link{GetDistance}()},
\code{\link{GetWeekPredCI}()},
\code{\link{LoadData}()},
\code{\link{RCTRecruit-package}},
\code{\link{Time2Nsubjects}()},
\code{\link{gripsYR1}},
\code{\link{gripsYR2}},
\code{\link{plot.RCTPredCI}()}
}
\concept{Links}
\keyword{datasets}
