% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{plot.RCTPredCI}
\alias{plot.RCTPredCI}
\title{Plots RCTPredCI object}
\usage{
\method{plot}{RCTPredCI}(x, yMax = NULL, includeYR2 = FALSE, xGrid = TRUE, yGrid = TRUE, ...)
}
\arguments{
\item{x}{RCTPredCI object}

\item{yMax}{It sets upper value for the y-axis. It is useful when several
figures are plotted together}

\item{includeYR2}{Whether to plot predictions for 104 weeks. By default (default = FALSE)
it only plots the first 52 weeks.}

\item{xGrid}{Whether to plot vertical grid lines (default = TRUE)}

\item{yGrid}{Whether to plot horizontal grid lines (default = TRUE)}

\item{...}{Additional arguments to be passed as in \code{\link[base:plot]{base::plot()}} function}
}
\description{
Plots RCTPredCI object
}
\seealso{
Other Links: 
\code{\link{GetDistance}()},
\code{\link{GetWeekPredCI}()},
\code{\link{LoadData}()},
\code{\link{RCTRecruit-package}},
\code{\link{Time2Nsubjects}()},
\code{\link{gripsYR1}},
\code{\link{gripsYR2}},
\code{\link{gripsYR2Weekly}}
}
\concept{Links}
