% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-RCTRecruit.R
\docType{package}
\name{RCTRecruit-package}
\alias{RCTRecruit}
\alias{RCTRecruit-package}
\title{RCTRecruit: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Ioannis Malagaris \email{iomalaga@utmb.edu}

}
\keyword{internal}
