% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simDistance}
\alias{simDistance}
\title{Function: Calculate CI of Euclidean distance of predicted recruitment with
actual recruitment}
\usage{
simDistance(target, fillGaps = FALSE, nSim = 10000L)
}
\arguments{
\item{target}{A vector with the actual recruitment by week}

\item{fillGaps}{Whether to fill gaps in the data}

\item{nSim}{Number of simulations to run}
}
\value{
A list with two elements. The first element \code{dist} is a numeric
vector with length equal to \code{nSim} containing the simulated Euclidean
distance. The second \code{CI} shows the median and the 95\%CI Euclidean
distance.
}
\description{
Function: Calculate CI of Euclidean distance of predicted recruitment with
actual recruitment
}
\examples{
LoadData(gripsIM, ScreenDt, Enrolled)
target <- days2weeks(gripsYR2$ScreenDt, gripsYR2$Enrolled)$enrolled
res <- simDistance(target)
}
