% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{time2Nsubjects}
\alias{time2Nsubjects}
\title{Function: Simulate number of weeks needed to recruit a given number of
subjects}
\usage{
time2Nsubjects(
  nSub = 50,
  nSim = 10000,
  fillGaps = FALSE,
  cauchyWt = FALSE,
  coeff = 1
)
}
\arguments{
\item{nSub}{Number of subjects to recruit (default = 50)}

\item{nSim}{Number of simulations to run (default = 1e4)}

\item{fillGaps}{Whether to fill gaps in the data (default = FALSE)}

\item{cauchyWt}{Whether to use Cauchy weights for sampling (default = FALSE).
If FALSE, binomial weights are used.}

\item{coeff}{A coefficient to apply to the recruitment rate (default = 1)}
}
\value{
A list with two elements. The first element \code{weeks} is an integer
vector with length equal to \code{nSim} containing the results of the
simulation. The second \code{CI} shows the median and the 95\%CI.
}
\description{
Function: Simulate number of weeks needed to recruit a given number of
subjects
}
\examples{
LoadData(gripsIM, ScreenDt, Enrolled)
res <- time2Nsubjects()
}
