% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{GetDistance}
\alias{GetDistance}
\title{Euclidean distance between prediction and actual recruitment}
\usage{
GetDistance(
  target,
  nSim = 10000L,
  fillGaps = FALSE,
  cauchyWt = FALSE,
  coeff = 1
)
}
\arguments{
\item{target}{A vector with the actual recruitment by week}

\item{nSim}{Number of simulations to run (default = 1e4L)}

\item{fillGaps}{Whether to fill gaps in the data (default = FALSE)}

\item{cauchyWt}{Whether to use Cauchy weights for sampling.
If FALSE (default),\if{html}{\out{<br>}}
binomial weights are used.}

\item{coeff}{A coefficient to apply to the recruitment rate (default = 1)}
}
\value{
An object of \code{RCTDist} class with four elements.
\enumerate{
\item \code{dist}:
A numeric vector with length equal to \code{nSim} containing the simulated\if{html}{\out{<br>}}
Euclidean distance.
\item \code{CI}: A numeric vector with the median and the 95\% CI Euclidean distance.
\item \code{call.}: The call (deparsed) that created the object
\item \code{cargs}: A list with the arguments of the call that created the object\if{html}{\out{<br>}}
including the default arguments
}
}
\description{
Euclidean distance between prediction and actual recruitment
}
\examples{
LoadData(gripsYR1, ScreenDt, Enrolled)
(res <- GetDistance(gripsYR2Weekly$enrolled))
str(res)
}
\seealso{
Other Links: 
\code{\link{GetWeekPredCI}()},
\code{\link{LoadData}()},
\code{\link{RCTRecruit-package}},
\code{\link{Time2Nsubjects}()},
\code{\link{gripsYR1}},
\code{\link{gripsYR2}},
\code{\link{gripsYR2Weekly}}
}
\concept{Links}
