% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{GetDistance}
\alias{GetDistance}
\title{Euclidean distance between prediction and actual recruitment}
\usage{
GetDistance(
  target,
  nSim = 10000L,
  fillGaps = FALSE,
  cauchyWt = FALSE,
  coeff = 1
)
}
\arguments{
\item{target}{A vector with the actual recruitment by week}

\item{nSim}{Number of simulations to run (default = 1e4)}

\item{fillGaps}{Whether to fill gaps in the data (default = FALSE)}

\item{cauchyWt}{Whether to use Cauchy weights for sampling (default = FALSE).
If FALSE, binomial weights are used.}

\item{coeff}{A coefficient to apply to the recruitment rate (default = 1)}
}
\value{
A list with two elements. The first element \code{dist} is a numeric
vector with length equal to \code{nSim} containing the simulated Euclidean
distance. The second \code{CI} shows the median and the 95\%CI Euclidean
distance.
}
\description{
Euclidean distance between prediction and actual recruitment
}
\examples{
LoadData(gripsYR1, ScreenDt, Enrolled)
res <- GetDistance(gripsWeeklyYR2$enrolled)
str(res)
}
\seealso{
Other RCTRecruit functions: 
\code{\link{GetWeekPredCI}()},
\code{\link{LoadData}()},
\code{\link{Time2Nsubjects}()}
}
\concept{RCTRecruit functions}
% f3
