% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDistance}
\alias{getDistance}
\title{Function: Calculate CI of Euclidean distance of predicted recruitment with
actual recruitment}
\usage{
getDistance(
  target,
  nSim = 10000,
  fillGaps = FALSE,
  cauchyWt = FALSE,
  coeff = 1
)
}
\arguments{
\item{target}{A vector with the actual recruitment by week}

\item{nSim}{Number of simulations to run}

\item{fillGaps}{Whether to fill gaps in the data}

\item{cauchyWt}{Whether to use Cauchy weights for sampling (default = FALSE).
If FALSE, binomial weights are used.}

\item{coeff}{A coefficient to apply to the recruitment rate (default = 1)}
}
\value{
A list with two elements. The first element \code{dist} is a numeric
vector with length equal to \code{nSim} containing the simulated Euclidean
distance. The second \code{CI} shows the median and the 95\%CI Euclidean
distance.
}
\description{
Function: Calculate CI of Euclidean distance of predicted recruitment with
actual recruitment
}
\examples{
LoadData(gripsIM, ScreenDt, Enrolled)
target <- days2weeks(gripsYR2$ScreenDt, gripsYR2$Enrolled)$enrolled
res <- getDistance(target)
}
