% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{GetWeekPredCI}
\alias{GetWeekPredCI}
\title{Calculate median recruitment with 95\% CI for the next 104 weeks (two years)}
\usage{
GetWeekPredCI(nSim = 10000L, fillGaps = FALSE, cauchyWt = FALSE, coeff = 1)
}
\arguments{
\item{nSim}{Number of simulations to run (default = 1e4L)}

\item{fillGaps}{Whether to fill gaps in the data (default = FALSE)}

\item{cauchyWt}{Whether to use Cauchy weights for sampling.
If FALSE (default),\if{html}{\out{<br>}}
binomial weights are used.}

\item{coeff}{A coefficient to apply to the recruitment rate (default = 1)}
}
\value{
An object of \code{RCTPredCI} class with 5 elements.
\enumerate{
\item \code{predCI}: An 104x3 matrix with the 2.5\%, 50\% and 97.5\% weekly percentiles
\item \code{plot(yMax = NULL, Title = NULL)}:\if{html}{\out{<br>}}
Function which plots the results. It accepts the following arguments:
\itemize{
\item \code{yMax} sets the high limit of the y-axis
\item \code{Title} sets the main title for the plot
}
\item \code{pargs}:\if{html}{\out{<br>}}
An environment which contains objects and functions used to construct\if{html}{\out{<br>}}
the plot. Additional plot configuration to what the \code{plot()} function\if{html}{\out{<br>}}
currently supports, can be achieved by modifying those objects
\item \code{call.}: The call (deparsed) that created the object
\item \code{cargs}: A list with the arguments of the call that created the object\if{html}{\out{<br>}}
including the default arguments
}
}
\description{
Calculate median recruitment with 95\% CI for the next 104 weeks (two years)
}
\examples{
LoadData(gripsYR1, ScreenDt, Enrolled)
(res <- GetWeekPredCI(fillGaps = TRUE, coeff = 1.5))
scenarios <- list(
  sc1 = GetWeekPredCI(),
  sc2 = GetWeekPredCI(cauchyWt = TRUE),
  sc3 = GetWeekPredCI(fillGaps = TRUE),
  sc4 = GetWeekPredCI(fillGaps = TRUE, coeff = 1.2)
)
maxY <- sapply(scenarios, \(x) x$pargs$maxY) |> max()
defaultGraphicParams <- par(no.readonly = TRUE)
graphics::par(mfrow = c(2, 2), oma = c(0, 1, 0, 7), mar = c(4, 4, 3, 1))
for (x in scenarios) x$plot(yMax = maxY, Title = x$call.)
do.call(par, defaultGraphicParams)
}
\seealso{
Other Links: 
\code{\link{GetDistance}()},
\code{\link{LoadData}()},
\code{\link{RCTRecruit-package}},
\code{\link{Time2Nsubjects}()},
\code{\link{gripsYR1}},
\code{\link{gripsYR2}},
\code{\link{gripsYR2Weekly}}
}
\concept{Links}
