% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{GetWeekPredCI}
\alias{GetWeekPredCI}
\title{Calculate median recruitment with CI for the next 52 weeks}
\usage{
GetWeekPredCI(nSim = 10000L, fillGaps = FALSE, cauchyWt = FALSE, coeff = 1)
}
\arguments{
\item{nSim}{Number of simulations to run (default = 1e4)}

\item{fillGaps}{Whether to fill gaps in the data (default = FALSE)}

\item{cauchyWt}{Whether to use Cauchy weights for sampling (default = FALSE).
If FALSE, binomial weights are used.}

\item{coeff}{A coefficient to apply to the recruitment rate (default = 1)}
}
\value{
An 52x3 matrix with the 2.5\%, 50\% and 97.5\% percentiles for each week
}
\description{
Calculate median recruitment with CI for the next 52 weeks
}
\examples{
LoadData(gripsYR1, ScreenDt, Enrolled)
res <- GetWeekPredCI(fillGaps = TRUE, coeff = 1.5)
str(res)
}
