% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getWeeksPredCI}
\alias{getWeeksPredCI}
\title{Function: Calculate median recruitment with CI for the next 52 weeks}
\usage{
getWeeksPredCI(nSim = 10000L, fill_gaps = FALSE)
}
\arguments{
\item{nSim}{Number of simulations to run}

\item{fill_gaps}{Whether to fill gaps in the data}
}
\value{
A 52x3 matrix with the 2.5\%, 50\% and 97.5\% percentiles for each week
}
\description{
Function: Calculate median recruitment with CI for the next 52 weeks
}
\examples{
LoadData(gripsIM, ScreenDt, Enrolled)
getWeeksPredCI()
}
