% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripsIM.R
\docType{data}
\name{gripsIM}
\alias{gripsIM}
\title{Data: Recruitment data for the GRIPS study}
\format{
A data frame with
\strong{159} observations of
\strong{2} variables:
\tabular{llll}{
[,1] \tab ScreenDt \tab \code{character} \tab Calendar date of the screening process for recruitment in the study\cr
[,2] \tab Enrolled \tab \code{integer} \tab Number of new subjects enrolled in the study
}
}
\source{
Villasante-Tezanos A, Kuo Y, Kurinec C, Li Y, Yu X (2024). "A
non-parametric approach to predict the recruitment for randomized
clinical trials: an example in elderly inpatient settings." \emph{BMC
medical research methodology}, \emph{24}, 189. ISSN 1471-2288,
\url{https://pubmed.ncbi.nlm.nih.gov/39210285/}.
}
\usage{
gripsIM
}
\description{
Modified recruitment data from the GRIPS study for the first year the study
was conducted. The data includes the number of participants recruited each
for each calendar date the recruitment was active.
}
\references{
\enumerate{
\item Villasante-Tezanos A, Kuo Y, Kurinec C, Li Y, Yu X (2024). "A
non-parametric approach to predict the recruitment for randomized
clinical trials: an example in elderly inpatient settings." \emph{BMC
medical research methodology}, \emph{24}, 189. ISSN 1471-2288,
\url{https://pubmed.ncbi.nlm.nih.gov/39210285/}.
\item Gajewski BJ, Simon SD, Carlson SE (2008). "Predicting accrual in
clinical trials with Bayesian posterior predictive distributions."
\emph{Statistics in medicine}, \emph{27}, 2328-40. ISSN 0277-6715,
\url{https://pubmed.ncbi.nlm.nih.gov/17979152/}.
\item Jiang Y, Simon S, Mayo MS, Gajewski BJ (2015). "Modeling and validating
Bayesian accrual models on clinical data and simulations using adaptive
priors." \emph{Statistics in medicine}, \emph{34}, 613-29. ISSN 1097-0258,
\url{https://pubmed.ncbi.nlm.nih.gov/25376910/}.
\item Jiang Y, Simon S, Mayo MS, Gajewski BJ (2015). "Modeling and validating
Bayesian accrual models on clinical data and simulations using adaptive
priors." \emph{Statistics in medicine}, \emph{34}, 613-29. ISSN 1097-0258,
\url{https://pubmed.ncbi.nlm.nih.gov/25376910/}.
}
}
\keyword{datasets}
