% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripsIM.R
\docType{data}
\name{gripsIM}
\alias{gripsIM}
\title{Recruitment data for the GRIPS study}
\format{
A data frame with 159 observations of 2 variables:

\tabular{llll}{
[,1] \tab ScreenDt \tab \code{character} \tab Calendar Date of the Screening\cr
[,2] \tab Enrolled \tab \code{integer} \tab Number of new subjects enrolled in the study
}
}
\source{
\url{https://github.com/imalagaris/RCTRecruit/tree/main/data}
}
\usage{
gripsIM
}
\description{
Modified recruitment data from the GRIPS study for the first year the study
was conducted. The data includes the number of participants recruited each
for each calendar date the recruitment was active.
}
\references{
References
}
\keyword{datasets}
