% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripsIM.r
\docType{data}
\name{gripsIM}
\alias{gripsIM}
\title{Recruitment data for the GRIPS study}
\format{
A data frame with
\strong{159} observations of
\strong{2} variables:
\tabular{llll}{
[,1] \tab ScreenDt \tab \code{character} \tab Calendar date of the screening process for recruitment in the study\cr
[,2] \tab Enrolled \tab \code{integer} \tab Number of new subjects enrolled in the study
}
}
\source{
Villasante-Tezanos A, Kuo Y, Kurinec C, Li Y, Yu X (2024). "A
non-parametric approach to predict the recruitment for randomized
clinical trials: an example in elderly inpatient settings." \emph{BMC
medical research methodology}, \emph{24}, 189. ISSN 1471-2288,
doi:10.1186/s12874-024-02314-2
\url{https://doi.org/10.1186/s12874-024-02314-2},
\url{http://www.ncbi.nlm.nih.gov/pubmed/39210285\%20http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=PMC11363376}.
}
\usage{
gripsIM
}
\description{
Modified recruitment data from the GRIPS study for the first year the study
was conducted. The data includes the number of participants recruited each
for each calendar date the recruitment was active.
}
\references{
\enumerate{
\item Villasante-Tezanos A, Kuo Y, Kurinec C, Li Y, Yu X (2024). "A
non-parametric approach to predict the recruitment for randomized
clinical trials: an example in elderly inpatient settings." \emph{BMC
medical research methodology}, \emph{24}, 189. ISSN 1471-2288,
doi:10.1186/s12874-024-02314-2
\url{https://doi.org/10.1186/s12874-024-02314-2},
\url{http://www.ncbi.nlm.nih.gov/pubmed/39210285\%20http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=PMC11363376}.
\item Villasante-Tezanos A, Kuo Y, Kurinec C, Li Y, Yu X (2024). "A
non-parametric approach to predict the recruitment for randomized
clinical trials: an example in elderly inpatient settings." \emph{BMC
medical research methodology}, \emph{24}, 189. ISSN 1471-2288,
doi:10.1186/s12874-024-02314-2
\url{https://doi.org/10.1186/s12874-024-02314-2},
\url{http://www.ncbi.nlm.nih.gov/pubmed/39210285\%20http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=PMC11363376}.
}
}
\keyword{datasets}
