% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gripsWeeklyYR2}
\alias{gripsWeeklyYR2}
\title{Data: Recruitment data for the GRIPS study}
\format{
A data frame with
\strong{52} observations of
\strong{5} variables:
\tabular{llll}{
[,1] \tab week \tab \code{double} \tab Calendar week\cr
[,2] \tab year \tab \code{double} \tab Calendar year\cr
[,3] \tab enrolled \tab \code{integer} \tab Number of people enrolled that week\cr
[,4] \tab holiday \tab \code{integer} \tab Number of federal holidays that week\cr
[,5] \tab cnt \tab \code{integer} \tab Number of days in that week when recruitment was active
}
}
\source{
Villasante-Tezanos A, Kuo Y, Kurinec C, Li Y, Yu X (2024). "A
non-parametric approach to predict the recruitment for randomized
clinical trials: an example in elderly inpatient settings." \emph{BMC
medical research methodology}, \emph{24}, 189. ISSN 1471-2288,
\url{https://pubmed.ncbi.nlm.nih.gov/39210285/}.
}
\usage{
gripsWeeklyYR2
}
\description{
Modified recruitment data aggregated by week from the second year of the
GRIPS study
}
\keyword{datasets}
