% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gripsYR1}
\alias{gripsYR1}
\title{Daily recruitment data for the 1st year of the GRIPS study}
\format{
A data frame with
\strong{159} observations of
\strong{2} variables:\tabular{llll}{
    \tab  \tab  \tab  \cr
   [,1] \tab ScreenDt \tab \code{character} \tab Calendar date of the screening process for recruitment in the study \cr
   [,2] \tab Enrolled \tab \code{integer} \tab Number of new subjects enrolled in the study on that date \cr
}
}
\source{
Villasante-Tezanos A, Kuo Y, Kurinec C, Li Y, Yu X (2024). "A
non-parametric approach to predict the recruitment for randomized
clinical trials: an example in elderly inpatient settings." \emph{BMC
medical research methodology}, \emph{24}, 189. ISSN 1471-2288,
\url{https://pubmed.ncbi.nlm.nih.gov/39210285/}.
}
\usage{
gripsYR1
}
\description{
Modified recruitment data from the GRIPS study for the first year the study
was conducted. The data includes the number of participants recruited each
for each calendar date the recruitment was active.
}
\keyword{datasets}
% d1
