% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{LoadData}
\alias{LoadData}
\title{Load recruitment data}
\usage{
LoadData(data, date, enrolled)
}
\arguments{
\item{data}{Main dataset}

\item{date}{Date column}

\item{enrolled}{Enrolled column}
}
\description{
Load recruitment data
}
\examples{
LoadData(gripsYR1, ScreenDt, Enrolled)
LoadData(gripsYR1, "ScreenDt", "Enrolled")
gripsYR1 |> LoadData(ScreenDt, Enrolled)
}
\seealso{
Other Links: 
\code{\link{GetDistance}()},
\code{\link{GetWeekPredCI}()},
\code{\link{RCTRecruit-package}},
\code{\link{Time2Nsubjects}()},
\code{\link{gripsWeeklyYR2}},
\code{\link{gripsYR1}},
\code{\link{gripsYR2}}
}
\concept{Links}
% f1
