% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LoadData}
\alias{LoadData}
\title{Function: Load recruitment data}
\usage{
LoadData(data, date, enrolled)
}
\arguments{
\item{data}{Main dataset}

\item{date}{Date column}

\item{enrolled}{Enrolled column}
}
\description{
Function: Load recruitment data
}
\examples{
LoadData(gripsIM, ScreenDt, Enrolled)
}
