% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-RCTRecruit.R
\docType{package}
\name{RCTRecruit-package}
\alias{RCTRecruit}
\alias{RCTRecruit-package}
\title{RCTRecruit: Non-parametric Recruitment Prediction for Randomized Clinical Trials}
\description{
Accurate prediction of subject recruitment for randomized clinical trials (RCT) remains an ongoing challenge. Previous prediction models rely on parametric assumptions. We present functions for non-parametric RCT recruitment prediction under several scenarios.
}
\details{
\tabular{ll}{
   Package: \tab RCTRecruit \cr
   Type: \tab Package \cr
   Version: \tab 0.1.0 \cr
   Date: \tab 2024-11-02 \cr
}
}
\references{
\enumerate{
\item \verb{r addRef("Villasante2024")}
\item \verb{r addRef("Gajewski2008")}
\item \verb{r addRef("Jiang2015")}
}
}
\seealso{
\verb{r addRef("accrual-package")}
}
\author{
\strong{Maintainer}: Ioannis Malagaris \email{iomalaga@utmb.edu} (\href{https://orcid.org/0000-0001-5126-2068}{ORCID})

Authors:
\itemize{
  \item Alejandro Villasante-Tezanos \email{alvillas@utmb.edu}
  \item Christopher Kurinec \email{chkurine@utmb.edu}
  \item Xiaoying Yu \email{xiyu@utmb.edu}
}

}
\keyword{internal}
