% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-RCTRecruit.R
\docType{package}
\name{RCTRecruit-package}
\alias{RCTRecruit}
\alias{RCTRecruit-package}
\title{RCTRecruit: Non-parametric Recruitment Prediction for Randomized ClinicalTrials}
\description{
Accurate prediction of subject recruitment for randomizedclinical trials (RCT) remains anongoing challenge. Previous predictionmodels rely on parametric assumptions. We present functions fornon-parametric RCT recruitment prediction under several scenarios.
}
\author{
\strong{Maintainer}: Ioannis Malagaris \email{iomalaga@utmb.edu} (\href{https://orcid.org/0000-0001-5126-2068}{ORCID})

Authors:
\itemize{
  \item Alejandro Villasante-Tezanos \email{alvillas@utmb.edu}
  \item Christopher Kurinec \email{chkurine@utmb.edu}
  \item Xiaoying Yu \email{xiyu@utmb.edu}
}

}
\keyword{internal}
