% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simAllWt}
\alias{simAllWt}
\title{Function: Simulate number of weeks needed to recruit a given number of
subjects}
\usage{
simAllWt(nSub = 50L, fill_gaps = FALSE, nSim = 10000L, startWK = 1L)
}
\arguments{
\item{nSub}{Number of subjects to recruit}

\item{fill_gaps}{Whether to fill gaps in the data}

\item{nSim}{Number of simulations to run}

\item{startWK}{The week of the year to start from}
}
\value{
A list with two elements. The first element \code{weeks} is an integer
vector with length equal to \code{nSim} containing the results of the
simulation. The second \code{CI} shows the median and the 95\%CI.
}
\description{
Function: Simulate number of weeks needed to recruit a given number of
subjects
}
\examples{
LoadData(gripsIM, ScreenDt, Enrolled)
simAllWt(50L)
}
